Make texture float extension optional
Safari will render to a float texture even though it reports that
it doesn't support the extension. Will need a new test for compatibility
to show an "unsupported browser" message. Also, Safari cannot read
pixels from a float texture, so will need to work around that to fully
support Safari.