Refactor sizing logic to be handled at top level of GL code
Since we want a vector view in addition to the raster view,
move resize code the top level and pass width and height to each view.
Also call the top level class GLManager to reflect the fact that it
handles all views plus the compute side of the WebGL calculation.